requirements.txt
mcp
fastmcp
anthropic
python-dotenv
requests
beautifulsoup4
pyjwt[crypto]
authlib
fastapi
uvicorn